chore(structure): Add .gitkeep to track of empty folders for gihub
GitHub does not show empty folders because Git tracks content, not the folders themselves. To push an "empty" folder to GitHub, you can include a placeholder file inside the folder to ensure it gets tracked. I added a file called .gitkeep so that the fodler is tracked inside of github